+++ pages/group/[id].tsx
@@ -1,13 +1,26 @@
-import { useRouter } from "next/router";
-import { ImageGrid } from "../../components/ImageGrid";
+import { GetServerSideProps, GetStaticProps } from "next";
+import { ImageGrid, ImageGridProps } from "../../components/ImageGrid";
+import { queryFnGetImageGroup } from "../../components/useGetImageGroup";
 
-export default function GroupPage() {
-  const router = useRouter();
-  const id = router.query.id as string;
+export const getServerSideProps: GetServerSideProps = async ({ params }) => {
+  const groupId = params?.id as string;
 
+  const data = await queryFnGetImageGroup({ queryKey: [groupId] });
+
+  const props: ImageGridProps = {
+    groupId,
+    initialData: data,
+  };
+
+  return {
+    props,
+  };
+};
+
+export default function GroupPage(props: ImageGridProps) {
   return (
     <div>
-      <ImageGrid groupId={id} />
+      <ImageGrid {...props} />
     </div>
   );
 }